Added cli_resful.c and moved system commands to cli_system.c.
