+++ components/iota_c/CMakeLists.txt
@@ -3,48 +3,70 @@ set(IOTA_SRC_DIR "iota_c/src")
 set(CRYPTO_SRCS "${IOTA_SRC_DIR}/crypto/iota_crypto.c")
 
 set(CORE_SRCS
-    "${IOTA_SRC_DIR}/core/seed.c"
     "${IOTA_SRC_DIR}/core/address.c"
     "${IOTA_SRC_DIR}/core/models/inputs/utxo_input.c"
+    "${IOTA_SRC_DIR}/core/models/outputs/byte_cost_config.c"
+    "${IOTA_SRC_DIR}/core/models/outputs/feat_blocks.c"
+    "${IOTA_SRC_DIR}/core/models/outputs/output_alias.c"
+    "${IOTA_SRC_DIR}/core/models/outputs/output_basic.c"
+    "${IOTA_SRC_DIR}/core/models/outputs/output_foundry.c"
+    "${IOTA_SRC_DIR}/core/models/outputs/output_nft.c"
+    "${IOTA_SRC_DIR}/core/models/outputs/native_tokens.c"
     "${IOTA_SRC_DIR}/core/models/outputs/outputs.c"
-    "${IOTA_SRC_DIR}/core/models/payloads/indexation.c"
+    "${IOTA_SRC_DIR}/core/models/outputs/storage_deposit.c"
+    "${IOTA_SRC_DIR}/core/models/outputs/unlock_conditions.c"
     "${IOTA_SRC_DIR}/core/models/payloads/transaction.c"
+    "${IOTA_SRC_DIR}/core/models/payloads/tagged_data.c"
+    "${IOTA_SRC_DIR}/core/models/payloads/milestone.c"
     "${IOTA_SRC_DIR}/core/models/message.c"
+    "${IOTA_SRC_DIR}/core/models/signing.c"
     "${IOTA_SRC_DIR}/core/models/unlock_block.c"
-    "${IOTA_SRC_DIR}/core/utils/bech32.c"
-    "${IOTA_SRC_DIR}/core/utils/byte_buffer.c"
     "${IOTA_SRC_DIR}/core/utils/iota_str.c"
-    "${IOTA_SRC_DIR}/core/utils/slip10.c")
+    "${IOTA_SRC_DIR}/core/utils/byte_buffer.c"
+    "${IOTA_SRC_DIR}/core/utils/slip10.c"
+    "${IOTA_SRC_DIR}/core/utils/bech32.c"
+    "${IOTA_SRC_DIR}/core/utils/uint256.c"
+    "${IOTA_SRC_DIR}/core/utils/macros.h")
 
 set(CLIENT_SRCS
     "${IOTA_SRC_DIR}/client/network/http_esp32.c"
     "${IOTA_SRC_DIR}/client/network/mqtt/mqtt_esp32.c"
-    "${IOTA_SRC_DIR}/client/api/json_utils.c"
-    "${IOTA_SRC_DIR}/client/api/message_builder.c"
-    "${IOTA_SRC_DIR}/client/api/message.c"
-    "${IOTA_SRC_DIR}/client/api/v1/find_message.c"
-    "${IOTA_SRC_DIR}/client/api/v1/get_balance.c"
-    "${IOTA_SRC_DIR}/client/api/v1/get_health.c"
-    "${IOTA_SRC_DIR}/client/api/v1/get_message_children.c"
-    "${IOTA_SRC_DIR}/client/api/v1/get_message_metadata.c"
-    "${IOTA_SRC_DIR}/client/api/v1/get_message.c"
-    "${IOTA_SRC_DIR}/client/api/v1/get_node_info.c"
-    "${IOTA_SRC_DIR}/client/api/v1/get_output.c"
-    "${IOTA_SRC_DIR}/client/api/v1/get_outputs_from_address.c"
-    "${IOTA_SRC_DIR}/client/api/v1/get_tips.c"
-    "${IOTA_SRC_DIR}/client/api/v1/response_error.c"
-    "${IOTA_SRC_DIR}/client/api/v1/send_message.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/common.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/json_utils.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/message.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/unlock_blocks.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/inputs/inputs.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/outputs/feat_blocks.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/outputs/native_tokens.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/outputs/output_alias.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/outputs/output_basic.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/outputs/output_foundry.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/outputs/output_nft.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/outputs/outputs.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/outputs/unlock_conditions.c"
+    "${IOTA_SRC_DIR}/client/api/json_parser/payloads/payloads.c"
+    "${IOTA_SRC_DIR}/client/api/restful/get_node_info.c"
+    "${IOTA_SRC_DIR}/client/api/restful/response_error.c"
+    "${IOTA_SRC_DIR}/client/api/restful/get_tips.c"
+    "${IOTA_SRC_DIR}/client/api/restful/get_output.c"
+    "${IOTA_SRC_DIR}/client/api/restful/get_health.c"
+    "${IOTA_SRC_DIR}/client/api/restful/get_message_children.c"
+    "${IOTA_SRC_DIR}/client/api/restful/get_message_metadata.c"
+    "${IOTA_SRC_DIR}/client/api/restful/get_message.c"
+    "${IOTA_SRC_DIR}/client/api/restful/get_milestone.c"
+    "${IOTA_SRC_DIR}/client/api/restful/send_message.c"
+    "${IOTA_SRC_DIR}/client/api/restful/send_tagged_data.c"
+    "${IOTA_SRC_DIR}/client/api/restful/faucet_enqueue.c"
+    "${IOTA_SRC_DIR}/client/api/restful/get_outputs_id.c"
+    "${IOTA_SRC_DIR}/client/api/restful/get_transaction_included_message.c"
     "${IOTA_SRC_DIR}/client/api/events/node_event.c"
-    "${IOTA_SRC_DIR}/client/api/events/sub_milestone_latest.c"
-    "${IOTA_SRC_DIR}/client/api/events/sub_milestones_confirmed.c"
+    "${IOTA_SRC_DIR}/client/api/events/sub_milestone_payload.c"
     "${IOTA_SRC_DIR}/client/api/events/sub_messages_metadata.c"
-    "${IOTA_SRC_DIR}/client/api/events/sub_serialized_output.c"
-    "${IOTA_SRC_DIR}/client/api/events/sub_outputs_payload.c")
+    "${IOTA_SRC_DIR}/client/api/events/sub_outputs_payload.c"
+    "${IOTA_SRC_DIR}/client/api/events/sub_serialized_output.c")
 
-set(WALLET_SRCS 
-    "${IOTA_SRC_DIR}/wallet/wallet.c"
-    "${IOTA_SRC_DIR}/wallet/bip39.c"
-)
+set(WALLET_SRCS "${IOTA_SRC_DIR}/wallet/wallet.c"
+                "${IOTA_SRC_DIR}/wallet/bip39.c")
 
 idf_component_register(
   SRCS
