+++ main/cli_restful.h
@@ -1,0 +1,14 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+/**
+ * @brief Set node enpoint for resful api calls
+ */
+void set_resftul_node_endpoint();
+
+/**
+ * @brief Register restful cli commands
+ */
+void register_restful_commands();
