+++ main/cli_system.c
@@ -1,0 +1,116 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+#include "cli_system.h"
+#include "esp_console.h"
+#include "esp_log.h"
+#include "esp_spi_flash.h"
+#include "esp_system.h"
+#include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
+
+static const char *TAG = "system";
+
+/* 'version' command */
+static int fn_get_version(int argc, char **argv) {
+  esp_chip_info_t info;
+  esp_chip_info(&info);
+  printf("IDF Version:%s\r\n", esp_get_idf_version());
+  printf("Chip info:\r\n");
+  printf("\tmodel:%s\r\n", CONFIG_IDF_TARGET);
+  printf("\tcores:%d\r\n", info.cores);
+  printf("\tfeature:%s%s%s%s%d%s\r\n", info.features & CHIP_FEATURE_WIFI_BGN ? ", 802.11bgn" : "",
+         info.features & CHIP_FEATURE_BLE ? ", BLE" : "", info.features & CHIP_FEATURE_BT ? ", BT" : "",
+         info.features & CHIP_FEATURE_EMB_FLASH ? ", Embedded-Flash:" : ", External-Flash:",
+         spi_flash_get_chip_size() / (1024 * 1024), " MB");
+  printf("\trevision number:%d\r\n", info.revision);
+  printf("Wallet version: v%s\n", APP_WALLET_VERSION);
+  return 0;
+}
+
+static void register_version() {
+  const esp_console_cmd_t cmd = {
+      .command = "version",
+      .help = "Show the esp32 and wallet versions",
+      .hint = NULL,
+      .func = &fn_get_version,
+  };
+  ESP_ERROR_CHECK(esp_console_cmd_register(&cmd));
+}
+
+/* 'restart' command */
+static int fn_restart(int argc, char **argv) {
+  ESP_LOGI(TAG, "Restarting");
+  esp_restart();
+}
+
+static void register_restart() {
+  const esp_console_cmd_t cmd = {
+      .command = "restart",
+      .help = "System reboot",
+      .hint = NULL,
+      .func = &fn_restart,
+  };
+  ESP_ERROR_CHECK(esp_console_cmd_register(&cmd));
+}
+
+/* 'free' command */
+static int fn_free_mem(int argc, char **argv) {
+  printf("%d\n", esp_get_free_heap_size());
+  return 0;
+}
+
+static void register_free() {
+  const esp_console_cmd_t cmd = {
+      .command = "free",
+      .help = "Get the size of available heap.",
+      .hint = NULL,
+      .func = &fn_free_mem,
+  };
+  ESP_ERROR_CHECK(esp_console_cmd_register(&cmd));
+}
+
+/* 'heap' command */
+static int fn_heap_size(int argc, char **argv) {
+  printf("heap info (SPI RAM): \n");
+  heap_caps_print_heap_info(MALLOC_CAP_SPIRAM);
+  printf("\nheap info (DEFAULT): \n");
+  heap_caps_print_heap_info(MALLOC_CAP_DEFAULT);
+  return 0;
+}
+
+static void register_heap() {
+  const esp_console_cmd_t heap_cmd = {
+      .command = "heap",
+      .help = "Get heap memory info",
+      .hint = NULL,
+      .func = &fn_heap_size,
+  };
+  ESP_ERROR_CHECK(esp_console_cmd_register(&heap_cmd));
+}
+
+/* 'stack' command */
+static int fn_stack_info(int argc, char **argv) {
+  printf("%u tasks are running on the system\n", uxTaskGetNumberOfTasks());
+  printf("Main stack size: %d, remaining %d bytes\n", CONFIG_MAIN_TASK_STACK_SIZE, uxTaskGetStackHighWaterMark(NULL));
+  return 0;
+}
+
+static void register_stack_info() {
+  const esp_console_cmd_t stack_info_cmd = {
+      .command = "stack",
+      .help = "Get system stack info",
+      .hint = NULL,
+      .func = &fn_stack_info,
+  };
+  ESP_ERROR_CHECK(esp_console_cmd_register(&stack_info_cmd));
+}
+
+void register_system_commands() {
+  // system info
+  register_free();
+  register_heap();
+  register_stack_info();
+  register_version();
+  register_restart();
+}
