--- main/sensor.h
@@ -1,7 +1,0 @@
-// Copyright 2021 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-void init_tempsensor();
-float get_temp();
